Relay and Retriever Subsystem has been defined but Relay is simply being
used as a sample test subsystem. The DriveTrain subsystem has been
changed as well so that it can accommodate for 3 transmissions for
controlling right and left parts of the robot. The retriever subsystem
has methods for controlling a double solenoid to release the retriever.
The retriever subsystem also has a relay to control the eject and
retrieve for balls.